完善jd订单拉取

+++ open-api/jd-api/src/main/java/com/qihang/jd/openApi/PullRequest.java
@@ -4,5 +4,5 @@ import lombok.Data;
 
 @Data
 public class PullRequest {
-    private Integer shopId;//店铺Id
+    private Long shopId;//店铺Id
 }
